// src/front/src/api/identity/models/structureResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { StructureRegion } from "./structureRegion";

export interface StructureResponse {
	id?: number;
	/** @nullable */
	code?: string | null;
	/** @nullable */
	name?: string | null;
	region?: StructureRegion;
	isActive?: boolean;
}